+++ contracts/tools/testing/IexecPocoBoostCompositeDelegate.sol
@@ -14,6 +14,4 @@ contract IexecPocoBoostCompositeDelegate is
     IexecPocoAccessorsDelegate,
     IexecPocoBoostAccessorsDelegate,
     IexecPocoBoostDelegate
-{
-
-}
+{}