+++ contracts/tools/testing/IexecPocoCompositeDelegate.sol
@@ -9,6 +9,4 @@ import {IexecPoco1Delegate} from "../../modules/delegates/IexecPoco1Delegate.sol
 /**
  * @notice This contract is dedicated to unit testing.
  */
-contract IexecPocoCompositeDelegate is IexecPoco1Delegate, IexecPocoAccessorsDelegate {
-
-}
+contract IexecPocoCompositeDelegate is IexecPoco1Delegate, IexecPocoAccessorsDelegate {}